Update alps.sh with vecreg
Based on recent publications (PMID: 37162692 and PMID: 36472803), it seems that ALPS is more robust when the tensor is transformed to the template through the FSL's vecreg function. Implemented this function in the pipeline (for now only when the default linear transformation to the template is performed, i.e., option -w 0 or not specified).