Update alps.sh
Added lines in the case of native space analysis (-t 0). 
If the specified ROIs correspond to the default JHU ROIs, then it will perform linear registration of the JHU FA template to the dti_FA in native space and apply the registration matrix to the JHU ROIs, which will be registered to the native space. 
Then these new ROIs in native space will be used for the computation of the ALPS.