Update alps.sh (-s1 option with ROI analysis in template space and user's custom ROIs)
Added if statements to perform registration of the FA/MD from native to template only if it does not exist. This is important for the scenario when the user wants to use the -s 1 option, i.e. already processed the images, but wants to calculate a new ALPS with new user-defined ROIs (previously would re-register the FA to template regardless, potentially leading to different results compared to the original run).